+++ server/src/authed.rs
@@ -843,7 +843,6 @@ pub async fn update_group_member(
                 auth.group_id,
                 message,
                 drop.item_id,
-                config.web_origin.clone(),
             );
         }
     }